requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cookiecutter==2.1.1
discord-py-interactions==5.0.0
discord-typings==0.5.1
emoji==2.2.0
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
